Add files via upload
Fixed "netplay" multiplayer gamepad to work correctly as expected. This was lost while going from a single webrtc/data connection for 1 joiner to multiple for up to 3 joiners. emulator.js and index.html needed updates afterword to fix both the gamepad inputs in general on joiners and specifically to fix the analog inputs from joiners once every other part of the gamepad was working. It all works now!