Add files via upload
Previous answers/offers in rooms are now erased when the Host starts a session for a room so that the joiners can join cleanly for Netplay -> Shared Console. 
Double tapping makes the joiner's screen fullscreen. 
Sound disabled on start of joining to fix issues where it would not work with iphones otherwise
Fixed controllers not working correctly on android. Previously they were navigating the system wide UI of the android. Now they are locked into the webpage (somehow. This was honestly a happy accident. I don't remember doing it)

Plan to change the function of the active profile icon to a joiner netplay option while being a joiner. So users can disconnect and other options.