updated indexn64netplay.html and core htmls
fixed an issue where you could not use the same room number without deleting the room folder or its contents. Instead on the start of hosting, the host overwrites the previous answers/offers with new ones, as it did not before, causing joiners to not connect correctly in previous used rooms. 

Uploaded n64 and psx specific htmls. They are the same as the indexn64netplay.html example, but n64 (exactly the same) and psx (psx core instead of the n64 core. Swapping out window.EJS_core        = "n64"; to the core you need is all the needs done if using a different system)