Add files via upload
set Game-profiles.html/profile.html to be dependent on profile.js to make updates less cumbersome. No longer need to switch user names out. profile.html is meant for the main media server package, but im uploading it as a fallback just in case here. Images for the profiles should be placed in /Profiles/
If hosting with this inside /emulator/ have a Profile folder with the images at /emulator/Profiles.